* GIST tracker added. RGB histogram added.

+++ examples/matlab2/tracker/Affine Sample Functions/warpimg.m
@@ -15,7 +15,7 @@ w = sz(2);  h = sz(1);  n = size(p,2);
 pos = reshape(cat(2, ones(h*w,1),x(:),y(:)) ...
     * [p(1,:) p(2,:); p(3:4,:) p(5:6,:)], [h,w,n,2]);
 
-wimg = squeeze(interp2(img, pos(:,:,:,1), pos(:,:,:,2)));
+wimg = squeeze(interp2(double(img), pos(:,:,:,1), pos(:,:,:,2)));
 
 wimg(find(isnan(wimg))) = 0;
